+++ src/app/components/common/Demo/DemoComponent.tsx
@@ -1,0 +1,10 @@
+import DemoFormContent from "./DemoFormContent";
+
+const DemoComponent = ({ className = "" }: { className?: string }) => {
+  return (
+    <div className={`w-full bg-white ${className}`}>
+      <DemoFormContent isFullPage={true} />
+    </div>
+  );
+};
+export default DemoComponent;
