+++ src/app/components/common/Demo/index.ts
@@ -1,0 +1,4 @@
+export { default as DemoTextArea } from "./DemoTextArea";
+export { default as DemoFormContent } from "./DemoFormContent";
+export { default as DemoComponent } from "./DemoComponent";
+export { default as DemoDialog } from "./DemoDialog";
