+++ src/app/components/ui/Agreement.tsx
@@ -4,14 +4,16 @@ import { Link } from "react-router-dom";
 const Agreement = () => {
   return (
     <>
-      By providing your contact information, you agree to our{" "}
+      <span className="text-grey">
+        By providing your contact information, you agree to our{" "}
+      </span>
       <Link
         to={APP_ROUTES.PUBLIC.TERMSCONDITION}
         className="text-primary underline"
       >
         Terms of Service
       </Link>{" "}
-      and{" "}
+      <span className="text-grey">and </span>
       <Link to={APP_ROUTES.PUBLIC.POLICY} className="text-primary underline">
         Privacy Policy
       </Link>
