--- src/app/features/auth/components/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as DemoTextArea } from "./DemoTextArea";
