+++ src/app/features/auth/index.ts
@@ -1,3 +1,2 @@
 export { default as LoginPage } from "./pages/LoginPage";
 export { default as RegisterPage } from "./pages/RegisterPage";
-export { default as DemoDialog } from "./pages/DemoDialog";
