+++ src/app/pages/index.ts
@@ -1,6 +1,7 @@
 export { default as NotFound } from "./NotFound";
 export { default as Unauthorized } from "./Unauthorized";
 export { default as Landing } from "./Landing";
+export { default as Demo } from "./Demo";
 
 // Product exports
 export { default as CRM } from "./products/crm/CRM";
